Extract a class to import a single content item
